Fixed TLE bug in countingSort()
Due to the nature of the countingSort function the i parameter of the for in radixSort function needs to be passed as argument also be the bitwise right operand, but, as i grow, the max_elem_array can possibly not have many digits as the bitwise right shift needs, so the bitwise operation wouldn't occur from this point but i you be growing as the for loop continues. In other words, if i isn't divided by 10, the for could possibly be infinite, and a runtime error would occur. Note that sorting an integer with more than 8 digits will result in the same runtime error.